fix: lazy loading some modules
